+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [@bpmn-io/cm-theme](https://github.com/bpmn-io/cm-theme) 
 
 ___Note:__ Yet to be released changes appear here._
 
+- `FIX`: restore feelers syntax highlighting ([#16](https://github.com/bpmn-io/cm-theme/pull/16))
+
 ## 0.2.0
 
 - `FEAT`: update theme with new colors and setup ([#5](https://github.com/bpmn-io/cm-theme/pull/5))